Revamp UI (#23)
This change revamps the single page format of the application into a multi-page format.

The new UI provides previews for both item selection and recording. The MainPage goes to the RecordingPage, which will go to the SavePage when successfully completing a recording. After saving or deleting the recording or if a failure occurs, users are sent back to the MainPage.